chore(haas): move unix domain events server to a module in a different file"
